# app/swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "User API",
      "description": "API pour gérer les utilisateurs"
    },
    "basePath": "/api",
    "schemes": ["http"],
    "securityDefinitions": {
      "Bearer": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header",
        "description": "Saisissez le token JWT dans le format : Bearer {token}"
      }
    },
    "security": [
      {
        "Bearer": []
      }
    ],
    "paths": {
      "/auth/register": {
        "post": {
          "summary": "Créer un nouvel utilisateur",
          "parameters": [
            {
              "name": "user",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Utilisateur enregistré avec succès"
            },
            "400": {
              "description": "Erreur de validation"
            }
          }
        }
      },
      "/auth/login": {
        "post": {
          "summary": "Connecter un utilisateur",
          "parameters": [
            {
              "name": "credentials",
              "in": "body",
              "required": true,
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": ["email", "password"]
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Connexion réussie avec JWT",
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string"
                  }
                }
              }
            },
            "400": {
              "description": "Erreur de validation ou mauvais identifiants"
            }
          }
        }
      },
      "/users": {
        "get": {
          "summary": "Récupérer tous les utilisateurs",
          "security": [
            {
              "Bearer": []
            }
          ],
          "responses": {
            "200": {
              "description": "Liste des utilisateurs",
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/definitions/User"
                }
              }
            }
          }
        }
      },
      "/users/{user_id}": {
        "get": {
          "summary": "Récupérer un utilisateur par son ID",
          "security": [
            {
              "Bearer": []
            }
          ],
          "parameters": [
            {
              "name": "user_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Détails de l'utilisateur",
              "schema": {
                "$ref": "#/definitions/User"
              }
            },
            "404": {
              "description": "Utilisateur non trouvé"
            }
          }
        },
        "put": {
          "summary": "Mettre à jour un utilisateur",
          "security": [
            {
              "Bearer": []
            }
          ],
          "parameters": [
            {
              "name": "user_id",
              "in": "path",
              "required": true,
              "type": "integer"
            },
            {
              "name": "user",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/User"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Utilisateur mis à jour avec succès"
            },
            "400": {
              "description": "Erreur de validation"
            },
            "404": {
              "description": "Utilisateur non trouvé"
            }
          }
        },
        "delete": {
          "summary": "Supprimer un utilisateur",
          "security": [
            {
              "Bearer": []
            }
          ],
          "parameters": [
            {
              "name": "user_id",
              "in": "path",
              "required": true,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "Utilisateur supprimé avec succès"
            },
            "404": {
              "description": "Utilisateur non trouvé"
            }
          }
        }
      }
    },
    "definitions": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": ["USER", "ADMIN"],
            "default": "USER"
          }
        }
      }
    }
  }
  